requirements.txt
aioamqp==0.13.0
amqp==2.6.1
asgiref==3.4.1
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
billiard==3.6.4.0
celery==4.3.0
certifi==2022.6.15
cffi==1.15.1
channels==2.2.0
channels-rabbitmq==1.1.5
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.4
daphne==2.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.4
django-betterforms==1.2
django-celery-beat==1.5.0
django-import-export==2.0.1
django-rosetta==0.9.3
django-timezone-field==4.2.3
djangorestframework==3.10.3
et-xmlfile==1.1.0
gevent==1.4.0
greenlet==1.1.2
html5lib==1.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.3
incremental==21.3.0
kombu==4.6.11
MarkupPy==1.14
msgpack==0.6.2
odfpy==1.4.1
openpyxl==3.0.10
pamqp==2.3.0
Pillow==8.4.0
pkg_resources==0.0.0
polib==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==0.10.0
pyOpenSSL==22.0.0
PyPDF2==2.8.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
reportlab==3.5.31
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tablib==0.14.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.10
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
zope.interface==5.4.0
